feedback

+++ codigo/controle/feedback.c
@@ -372,7 +372,7 @@
  * and open the template in the editor.
  */
 
-Cliente* clienteCodigoFeedback(int cod1, int cod2) {
+Cliente* clienteCodigoFeedback(float cod1, float cod2) {
     Cliente *Clientes;
     Clientes = listaClientes();
     Cliente *array = malloc((qtdCliente()) * sizeof (Cliente));
@@ -426,7 +426,7 @@ Cliente* clienteSexoFeedback(char sexo) {
     return array;
 }
 
-Filme* filmeCodigoFeedback(int cod1, int cod2) {
+Filme* filmeCodigoFeedback(float cod1, float cod2) {
     Filme *Filmes;
     Filmes = listaFilmes();
     Filme *array = malloc((qtdFilme()) * sizeof (Filme));
@@ -450,7 +450,7 @@ Filme* filmeCodigoFeedback(int cod1, int cod2) {
 
 //DAQUI PARA FRENTE NÃO FUNCIONA PQ TEM QUE PERSISTIR
 
-int qtdeLocacoesParaQueSePague(int cod, float precoTotal, float precoFilme) {
+int qtdeLocacoesParaQueSePague(float cod, float precoTotal, float precoFilme) {
     //Não entendi isso aqui não
     Locacao *Locacoess;
     float qtde;
@@ -618,7 +618,7 @@ Conta * contasReceberDataFeedback(char *data1, char *data2) {
     return array;
 }
 
-Conta * contasReceberCodigoFeedback(int cod1, int cod2) {
+Conta * contasReceberCodigoFeedback(float cod1, float cod2) {
     Conta *Contas;
     Contas = listaContas();
     Conta *array = malloc((qtdConta()) * sizeof (Conta));
@@ -682,7 +682,7 @@ array[j-1].codigo=-1;
     return array;
 }
 
-Conta * contasPagarCodigoFeedback(int cod1, int cod2) {
+Conta * contasPagarCodigoFeedback(float cod1, float cod2) {
     Conta *Contas;
     Contas = listaContas();
     Conta *array = malloc((qtdConta()) * sizeof (Conta));
